modify deploy.sh so it throws an error and halts if step 1 of the deploy fails

+++ deploy.sh
@@ -64,17 +64,25 @@ echo 'Deploying the remaining metadata'
 if [ "$CHECKONLY" = true ]
 then
   echoOut 'VALIDATING ONLY'
-  sfdx force:source:deploy --testlevel $TESTLEVEL --checkonly --targetusername $1 -p force-app/main/default -g -w 120
-  echoOut 're-enable original .forceignore'
-  rm -f .forceignore
-  mv .forceignore.orig .forceignore
-  echoOut 'Deploying ExperienceBundles'
-  sfdx force:source:deploy --testlevel NoTestRun --checkonly --targetusername $1 -p force-app/main/default/experiences -g -w 120
+  if sfdx force:source:deploy --testlevel $TESTLEVEL --checkonly --targetusername $1 -p force-app/main/default -g -w 120 ; then
+    echoOut 're-enable original .forceignore'
+    rm -f .forceignore
+    mv .forceignore.orig .forceignore
+    echoOut 'Deploying ExperienceBundles'
+    sfdx force:source:deploy --testlevel NoTestRun --checkonly --targetusername $1 -p force-app/main/default/experiences -g -w 120
+  else
+    echoOut 'Validation Fail'
+    exit 1
+  fi
 else
-  sfdx force:source:deploy --testlevel $TESTLEVEL --targetusername $1 -p force-app/main/default -g -w 120
-  echoOut 're-enable original .forceignore'
-  rm -f .forceignore
-  mv .forceignore.orig .forceignore
-  echoOut 'Deploying ExperienceBundles'
-  sfdx force:source:deploy --testlevel NoTestRun --targetusername $1 -p force-app/main/default/experiences -g -w 120
+  if sfdx force:source:deploy --testlevel $TESTLEVEL --targetusername $1 -p force-app/main/default -g -w 120 ; then
+    echoOut 're-enable original .forceignore'
+    rm -f .forceignore
+    mv .forceignore.orig .forceignore
+    echoOut 'Deploying ExperienceBundles'
+    sfdx force:source:deploy --testlevel NoTestRun --targetusername $1 -p force-app/main/default/experiences -g -w 120
+  else
+    echoOut 'Deploy Fail'
+    exit 1
+  fi
 fi
